+++ packages/design-system-angular/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.16.1-pr-95.4 (2023-10-11)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.16.1-pr-95.3 (2023-10-11)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.16.1-pr-95.2 (2023-09-29)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.16.1-pr-95.1 (2023-09-29)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.16.1-pr-95.0 (2023-09-29)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
 ## 1.16.0 (2023-09-22)
 
 **Note:** Version bump only for package @geovistory/design-system-angular